seedcase/requirements.txt
asgiref==3.5.2
Django==4.1.2
sqlparse==0.4.3
psycopg2-binary>=2.9
style==1.1.0
six==1.16.0
update==0.0.1
django-debug-toolbar==3.2.2
django-nested-inline==0.4.5
djangorestframework==3.12.4
requests>=2.18.4
django-createsuperuser
django-bootstrap-v5==1.0.11
pytz==2022.4
pytest-django==3.8.0
pytest-cov==2.8.1
drf-yasg==1.21.5
